Refactor user service into MVC
