+++ microservices/userService/src/routes/index.js
@@ -1,0 +1,3 @@
+const userRoutes = require('./user.route');
+
+module.exports = { userRoutes }
